# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FitPass HOPn API Documentation",
        "description": "API Documentation for the FitPass HOPn Platform",
        "contact": {
            "email": "your-email@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Main API Server"
        },
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
        "/api/v1/admin/stats/platform-overview": {
            "get": {
                "tags": [
                    "Super Admin - Analytics"
                ],
                "summary": "Get a high-level overview of the entire platform",
                "operationId": "f37165ef59a6f69d2a1d94afbb847ca4",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/stats/company-activity": {
            "get": {
                "tags": [
                    "Super Admin - Analytics"
                ],
                "summary": "Get a summary of activity for each company",
                "operationId": "d6faa06c675626e6565a092e72408b2d",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/stats/partner-performance": {
            "get": {
                "tags": [
                    "Super Admin - Analytics"
                ],
                "summary": "Get performance stats for all partners",
                "operationId": "fc8e85707516ce73d6dddf4f55174d74",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticate user and return API token",
                "description": "//",
                "operationId": "bae68d1c55329ecb6e41f48285840c8f",
                "requestBody": {
                    "description": "User credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "device_name"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "hr@fitpass.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    },
                                    "device_name": {
                                        "description": "Identifier for the client device/application",
                                        "type": "string",
                                        "example": "Chrome Browser"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Login successful."
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "1|Abcdefghijklmnopqrstuvwxyz123456"
                                                },
                                                "user": {
                                                    "$ref": "#/components/schemas/UserResource"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error or invalid credentials"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user (invalidate current token)",
                "description": "//",
                "operationId": "c5cdd87fdbf9d0f95dede875bd023dec",
                "responses": {
                    "200": {
                        "description": "Successfully logged out",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully logged out."
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/my-bookings": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Get a list of the authenticated user's bookings",
                "operationId": "be0a09d3e68c850f53805c5c24a87123",
                "responses": {
                    "200": {
                        "description": "A list of the user's bookings.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/BookingResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/classes/{classId}/book": {
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Book the authenticated user into a specific class",
                "operationId": "b1e125663579d851c32488238d894214",
                "parameters": [
                    {
                        "name": "classId",
                        "in": "path",
                        "description": "The ID of the class to book",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Booking created successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/BookingResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Class not found"
                    },
                    "409": {
                        "description": "Class is full or already booked"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "BookingResource": {
                "title": "Booking Resource",
                "description": "Represents a user's booking for a class.",
                "properties": {
                    "booking_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "confirmed",
                            "cancelled"
                        ],
                        "example": "confirmed"
                    },
                    "booked_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-30T14:30:00.000000Z"
                    },
                    "class": {
                        "$ref": "#/components/schemas/ClassModelResource"
                    },
                    "partner": {
                        "$ref": "#/components/schemas/PartnerResource"
                    }
                },
                "type": "object"
            },
...
            "ClassModelResource": {
                "title": "Class Model Resource",
                "description": "Represents a single class or session.",
                "properties": {
                    "class_id": {
                        "type": "integer",
                        "example": 101
                    },
                    "name": {
                        "type": "string",
                        "example": "Morning Yoga"
                    },
                    "description": {
                        "type": "string",
                        "example": "A refreshing yoga session to start your day."
                    },
                    "start_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-05T09:00:00.000000Z"
                    },
                    "end_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-05T10:00:00.000000Z"
                    },
                    "capacity": {
                        "type": "integer",
                        "example": 20
                    }
                },
                "type": "object"
            },
...
            "PartnerResource": {
                "title": "Partner Resource",
                "description": "Represents a partner gym, spa, or club.",
                "properties": {
                    "partner_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Global Fitness"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "gym",
                            "spa",
                            "club"
                        ],
                        "example": "gym"
                    },
                    "city": {
                        "type": "string",
                        "example": "Lima"
                    },
                    "address": {
                        "type": "string",
                        "example": "Av. Larco 123"
                    },
                    "location": {
                        "properties": {
                            "latitude": {
                                "type": "number",
                                "format": "float",
                                "example": -12.1218
                            },
                            "longitude": {
                                "type": "number",
                                "format": "float",
                                "example": -77.0311
                            }
                        },
                        "type": "object"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "approved",
                            "rejected"
                        ],
                        "example": "approved"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "title": "User Resource",
                "description": "User model representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "HR Admin"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "hr@fitpass.com"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "employee",
                            "hr_admin",
                            "super_admin"
                        ],
                        "example": "hr_admin"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "null",
                        "nullable": true
                    },
                    "company_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "membership_plan_id": {
                        "type": "integer",
                        "example": 3,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-26T11:04:40.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-26T11:04:40.000000Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}